requirements.txt
asgiref==3.5.2
dj-static==0.0.6
Django==4.1.3
django-stdimage==6.0.1
django-widget-tweaks==1.4.12
gunicorn==20.1.0
Pillow==9.3.0
sqlparse==0.4.3
static3==0.7.0
whitenoise==6.2.0
